% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkArguments.R
\name{checkArguments}
\alias{checkArguments}
\title{Check Arguments for CCM estimands}
\usage{
checkArguments(Y, T1, T2, M, data = NULL)
}
\arguments{
\item{data}{}
}
\value{
A dataframe
}
\description{
Check Arguments for CCM estimands
}
