% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decideOutput.R
\name{decideOutput}
\alias{decideOutput}
\title{Decide output and calculate confidence intervals}
\usage{
decideOutput(para.df, noInteraction = TRUE, boots = 1000, sigLevel = 0.05)
}
\arguments{
\item{para.df}{A data.frame of arguments generated using checkArguments().}

\item{noInteraction}{Logical, TRUE by default.  If TRUE, there is no expected interaction between the treatments and mediator.}

\item{boots}{Number of bootstrap iterations, must be larger than 1000, 1000 by default.}

\item{sigLevel}{The significance level, 0.05 by default.}
}
\value{

}
\description{
Decide output and calculate confidence intervals
}
