% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICAapp.R
\docType{data}
\name{ICAapp}
\alias{ICAapp}
\title{ICAapp data}
\format{
a data frame with 14 variables

\describe{
\item{respid}{Needs update}
\item{appr}{Needs update}
\item{dapprp}{Needs update}
\item{vote}{Needs update}
\item{nvotep}{Needs update}
\item{moral}{Needs update}
\item{immorp}{Needs update}
\item{retal}{Needs update}
\item{retalp}{Needs update}
\item{trt1}{Needs update}
\item{trt2}{Needs update}
\item{age}{Needs update}
\item{male}{Needs update}
\item{educ}{Needs update}
}
}
\usage{
ICAapp
}
\description{
description for ICAapp data
}
\keyword{datasets}
