% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmEstimation.R
\name{summary.ccmEstimation}
\alias{summary.ccmEstimation}
\title{Summary of ccmEstimation}
\usage{
\method{summary}{ccmEstimation}(object, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Summary of ccmEstimation
}
