% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInteraction.R
\name{checkInteraction}
\alias{checkInteraction}
\title{Check Interaction between mediator and treatments}
\usage{
checkInteraction(para.df, sigLevel = 0.05)
}
\arguments{
\item{para.df}{A data.frame of arguments generated using checkArguments().}

\item{sigLevel}{The significance level, 0.05 by default.}
}
\value{
Logical. 1 if the interaction between treatments and mediator exists.
}
\description{
Check Interaction between mediator and treatments
}
\examples{
\dontrun{
df = checkArguments('dapprp','trt1','trt2','immorp',testdata)
interaction = checkInteration(df)
}

}
