% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInteraction.R
\name{checkInteraction}
\alias{checkInteraction}
\title{A function that checks the interaction}
\usage{
checkInteraction(para.df, sigLevel = 0.05)
}
\arguments{
\item{sigLevel}{}
}
\value{
a logical value. 1 if interaction between mediator and treatment exists.
}
\description{
A function that checks the interaction
}
