% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decideOutput.R
\name{decideOutput}
\alias{decideOutput}
\title{Decide output and calculate confidence intervals}
\usage{
decideOutput(para.df, noInteraction = TRUE, boots = 1000, sigLevel = 0.05)
}
\arguments{
\item{sigLevel}{}
}
\value{

}
\description{
Decide output and calculate confidence intervals
}
\examples{

}
