% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCCM.R
\name{getCCM}
\alias{getCCM}
\title{Get CCM estimands}
\usage{
getCCM(
  Y,
  T1,
  T2,
  M,
  data = NULL,
  noInteraction = TRUE,
  sigLevel = 0.05,
  boots = 1000,
  finiteSample = FALSE
)
}
\arguments{
\item{finiteSample}{}
}
\value{

}
\description{
Get CCM estimands
}
\examples{

}
