% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmEstimation.R
\name{print.summary.ccmEstimation}
\alias{print.summary.ccmEstimation}
\title{Print summary tables}
\usage{
\method{print}{summary.ccmEstimation}(x, ...)
}
\arguments{
\item{...}{}
}
\value{
Print the summary tables
}
\description{
Print summary tables
}
\examples{
#' \dontrun{
df = checkArguments('dapprp','trt1','trt2','immorp',testdata)
summary(df)
}
}
