% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmEstimation.R
\name{print.summary.ccmEstimation}
\alias{print.summary.ccmEstimation}
\title{Print summary}
\usage{
\method{print}{summary.ccmEstimation}(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Print summary
}
